// src/features/laporan/pivot/types/pivot.type.ts
export interface PivotData {
  id: string;
  nip: string;
  nama: string;
  departemen: string;
  sub_departemen: string;
  pos: string;
  grade: string;
  skema: string;
  masuk: number;
  libur: number;
  ph: number;
  izin: number;
  alfa: number;
  sakit: number;
  cuti: number;
  terlambat: number;
}